mybatis 전환 및 완성

+++ src/BoardApp.java
@@ -1,296 +1,39 @@
-import java.sql.*;
+//import java.sql.*;
 import java.util.ArrayList;
 import java.util.Scanner;
 
 public class BoardApp {
-//    private Scanner scanner = new Scanner(System.in);
-//    private Connection conn;
-
-//    //Constructor
-//    public BoardApp() {
-//        try {
-//            //JDBC Driver 등록
-//            Class.forName("org.mariadb.jdbc.Driver");
-//
-//            //DB 연결하기
-//            conn = DriverManager.getConnection(
-//                    "jdbc:mariadb://127.0.0.1:3306/board_prj",
-//                    "root",
-//                    "12345");
-//        } catch (Exception e) {
-//            e.printStackTrace();
-//            exit();
-//        }
-//    }
-
-//    //Method
-//    public void list() {
-//        //타이틀 및 컬럼명 출력
-//        System.out.println();
-//        System.out.println("[게시물 목록]");
-//        System.out.println("-----------------------------------------------------------------------");
-//        System.out.printf("%-6s%-12s%-16s%-40s\n", "no", "writer", "date", "title");
-//        System.out.println("-----------------------------------------------------------------------");
-//
-
-
-//        //boads 테이블에서 게시물 정보를 가져와서 출력하기
-//        try {
-//            String sql = "" +
-//                    "SELECT bno, btitle, bcontent, bwriter, bdate " +
-//                    "FROM boards " +
-//                    "ORDER BY bno DESC";
-//            PreparedStatement pstmt = conn.prepareStatement(sql);
-//            ResultSet rs = pstmt.executeQuery();
-//            while (rs.next()) {
-//                Board board = new Board();
-//                board.setBno(rs.getInt("bno"));
-//                board.setBtitle(rs.getString("btitle"));
-//                board.setBcontent(rs.getString("bcontent"));
-//                board.setBwriter(rs.getString("bwriter"));
-//                board.setBdate(rs.getDate("bdate"));
-//                System.out.printf("%-6s%-12s%-16s%-40s\n",
-//                        board.getBno(),
-//                        board.getBwriter(),
-//                        board.getBdate(),
-//                        board.getBtitle());
-//
-//            }
-//            rs.close();
-//            pstmt.close();
-//        } catch (SQLException e) {
-//            e.printStackTrace();
-//            exit();
-//        }
-//
-//        //메인 메뉴 출력
-//        mainMenu();
-//    }
-
-//    public void mainMenu() {
-//        System.out.println();
-//        System.out.println("-----------------------------------------------------------------------");
-//        System.out.println("메인메뉴: 1.Create | 2.Read | 3.Clear | 4.Exit");
-//        System.out.print("메뉴선택: ");
-//        String menuNo = scanner.nextLine();
-//        System.out.println();
-//
-//        switch (menuNo) {
-//            case "1":
-//                create();
-//            case "2":
-//                read();
-//            case "3":
-//                clear();
-//            case "4":
-//                exit();
-//        }
-//    }
-
-//    public void create() {
-//        //입력 받기
-//        Board board = new Board();
-//        System.out.println("[새 게시물 입력]");
-//        System.out.print("제목: ");
-//        board.setBtitle(scanner.nextLine());
-//        System.out.print("내용: ");
-//        board.setBcontent(scanner.nextLine());
-//        System.out.print("글쓴이: ");
-//        board.setBwriter(scanner.nextLine());
-//
-//        //보조메뉴 출력
-//        System.out.println("-----------------------------------------------------------------------");
-//        System.out.println("보조메뉴: 1.Ok | 2.Cancel");
-//        System.out.print("메뉴선택: ");
-//        String menuNo = scanner.nextLine();
-//        if (menuNo.equals("1")) {
-//            //boards 테이블에 게시물 정보 저장
-//            try {
-//                String sql = "" +
-//                        "INSERT INTO boards (btitle, bcontent, bwriter, bdate) " +
-//                        "VALUES (?, ?, ?, now())";
-//                PreparedStatement pstmt = conn.prepareStatement(sql);
-//                pstmt.setString(1, board.getBtitle());
-//                pstmt.setString(2, board.getBcontent());
-//                pstmt.setString(3, board.getBwriter());
-//                pstmt.executeUpdate();
-//                pstmt.close();
-//            } catch (Exception e) {
-//                e.printStackTrace();
-//                exit();
-//            }
-//        }
-//
-//        //게시물 목록 출력
-//        list();
-//    }
-//
-//    public void read() {
-//        //입력 받기
-//        System.out.println("[게시물 읽기]");
-//        System.out.print("bno: ");
-//        int bno = Integer.parseInt(scanner.nextLine());
-//
-//        //boards 테이블에서 해당 게시물을 가져와 출력
-//        try {
-//            String sql = "" +
-//                    "SELECT bno, btitle, bcontent, bwriter, bdate " +
-//                    "FROM boards " +
-//                    "WHERE bno=?";
-//            PreparedStatement pstmt = conn.prepareStatement(sql);
-//            pstmt.setInt(1, bno);
-//            ResultSet rs = pstmt.executeQuery();
-//            if (rs.next()) {
-//                Board board = new Board();
-//                board.setBno(rs.getInt("bno"));
-//                board.setBtitle(rs.getString("btitle"));
-//                board.setBcontent(rs.getString("bcontent"));
-//                board.setBwriter(rs.getString("bwriter"));
-//                board.setBdate(rs.getDate("bdate"));
-//                System.out.println("#############");
-//                System.out.println("번호: " + board.getBno());
-//                System.out.println("제목: " + board.getBtitle());
-//                System.out.println("내용: " + board.getBcontent());
-//                System.out.println("쓴이: " + board.getBwriter());
-//                System.out.println("날짜: " + board.getBdate());
-//                //보조메뉴 출력
-//                System.out.println("-------------------------------------------------------------------");
-//                System.out.println("보조메뉴: 1.Update | 2.Delete | 3.List");
-//                System.out.print("메뉴선택: ");
-//                String menuNo = scanner.nextLine();
-//                System.out.println();
-//
-//                if (menuNo.equals("1")) {
-//                    update(board);
-//                } else if (menuNo.equals("2")) {
-//                    delete(board);
-//                }
-//            }
-//            rs.close();
-//            pstmt.close();
-//        } catch (Exception e) {
-//            e.printStackTrace();
-//            exit();
-//        }
-//
-//        //게시물 목록 출력
-//        list();
-//    }
-//
-//    public void update(Board board) {
-//        //수정 내용 입력 받기
-//        System.out.println("[수정 내용 입력]");
-//        System.out.print("제목: ");
-//        board.setBtitle(scanner.nextLine());
-//        System.out.print("내용: ");
-//        board.setBcontent(scanner.nextLine());
-//        System.out.print("글쓴이: ");
-//        board.setBwriter(scanner.nextLine());
-//
-//        //보조메뉴 출력
-//        System.out.println("-------------------------------------------------------------------");
-//        System.out.println("보조메뉴: 1.Ok | 2.Cancel");
-//        System.out.print("메뉴선택: ");
-//        String menuNo = scanner.nextLine();
-//        if (menuNo.equals("1")) {
-//            //boards 테이블에서 게시물 정보 수정
-//            try {
-//                String sql = "" +
-//                        "UPDATE boards SET btitle=?, bcontent=?, bwriter=? " +
-//                        "WHERE bno=?";
-//                PreparedStatement pstmt = conn.prepareStatement(sql);
-//                pstmt.setString(1, board.getBtitle());
-//                pstmt.setString(2, board.getBcontent());
-//                pstmt.setString(3, board.getBwriter());
-//                pstmt.setInt(4, board.getBno());
-//                pstmt.executeUpdate();
-//                pstmt.close();
-//            } catch (Exception e) {
-//                e.printStackTrace();
-//                exit();
-//            }
-//        }
-//
-//        //게시물 목록 출력
-//        list();
-//    }
-//
-//    public void delete(Board board) {
-//        //boards 테이블에 게시물 정보 삭제
-//        try {
-//            String sql = "DELETE FROM boards WHERE bno=?" +
-//                    "SET @COUNT=0" +
-//                    "UPDATE boards SET boards.bno = @COUNT:=@COUNT+1";
-//            PreparedStatement pstmt = conn.prepareStatement(sql);
-//            pstmt.setInt(1, board.getBno());
-//            pstmt.executeUpdate();
-//            pstmt.close();
-//        } catch (Exception e) {
-//            e.printStackTrace();
-//            exit();
-//        }
-//
-//        //게시물 목록 출력
-//        list();
-//    }
-//    public void clear() {
-//        System.out.println("[게시물 전체 삭제]");
-//        System.out.println("-------------------------------------------------------------------");
-//        System.out.println("보조메뉴: 1.Ok | 2.Cancel");
-//        System.out.print("메뉴선택: ");
-//        String menuNo = scanner.nextLine();
-//        if(menuNo.equals("1")) {
-//            //boards 테이블에 게시물 정보 전체 삭제
-//            try {
-//                String sql = "TRUNCATE TABLE boards";
-//                PreparedStatement pstmt = conn.prepareStatement(sql);
-//                pstmt.executeUpdate();
-//                pstmt.close();
-//            } catch (Exception e) {
-//                e.printStackTrace();
-//                exit();
-//            }
-//        }
-//
-//        //게시물 목록 출력
-//        list();
-//    }
-
-//    public void exit() {
-//        if(conn != null) {
-//            try {
-//                conn.close();
-//            } catch (SQLException e) {
-//            }
-//        }
-//        System.out.println("** 게시판 종료 **");
-//        System.exit(0);
-//    }
-
     public static void main(String[] args) {
-//        BoardApp boardExam = new BoardApp();
-//        boardExam.list();
 
+        //데이터베이스 기능들을 담기 위한 초기화
         DBUtil db = new DBUtil();
         Scanner scan = new Scanner(System.in);
 
+        //마이바티스 설정을 위한 초기화
         db.init();
-        
+
+        //기본적으로 계속 프로그램이 돌아가도록 반복문
         while (true){
+            //데이터베이스에서 게시물 목록 가져오기
             ArrayList<Board> boardList = db.getBoard();
+            //기본적인 인터페이스 초기화
             WebView wv = new WebView();
+            //게시물 전체 목록 표시
             wv.printBoard(boardList);
             System.out.println();
             System.out.println("-----------------------------------------------------------------------");
             System.out.println("메인메뉴: 1.Create | 2.Read | 3.Clear | 4.Exit");
             System.out.print("메뉴선택: ");
+            //사용자의 원하는 메뉴 읽기
             String cmd = scan.nextLine();
             System.out.println();
 
+            //4를 입력하면 프로그램 종료
             if (cmd.equals("4")){
                 System.out.println("** 게시판 종료 **");
                 break;
             } else if (cmd.equals("1")) {
+                //1을 입력하면 게시물 생성으로 이동
                 System.out.println("[새 게시물 입력]");
                 System.out.print("제목: ");
                 String title = scan.nextLine();
@@ -301,7 +44,9 @@ public class BoardApp {
                 System.out.println("-----------------------------------------------------------------------");
                 System.out.println("보조메뉴: 1.Ok | 2.Cancel");
                 System.out.print("메뉴선택: ");
+                //게시글 만드는걸 확정하는 보조메뉴
                 String menuNo = scan.nextLine();
+                //1을 입력받으면 입력받은 제목, 내용, 글쓴이의 정보로 데이터베이스 새 게시물 생성
                 if(menuNo.equals("1")){
                     try {
                         db.createBoard(title,content,writer);
@@ -311,15 +56,23 @@ public class BoardApp {
                     }
                 }
             } else if (cmd.equals("2")) {
+                //2를 입력하면 원하는 번호의 게시물을 상세보기
                 System.out.println("[게시물 읽기]");
                 System.out.print("bno: ");
+                //원하는 게시물의 번호 입력받기
                 int bno = Integer.parseInt(scan.nextLine());
+                //데이터베이스에서 원하는 번호의 게시물 정보 가져오기
+                Board board = db.readBoard(bno);
+                //웹뷰에 있는 인터페이스 불러와서 원하는 게시물 정보 출력
+                wv.printReadBoard(board);
+                System.out.println();
                 System.out.println("-------------------------------------------------------------------");
                 System.out.println("보조메뉴: 1.Update | 2.Delete | 3.List");
                 System.out.print("메뉴선택: ");
                 String menuNo = scan.nextLine();
                 System.out.println();
 
+                //1번 업데이트를 선택한다면 방금 읽은 게시물 수정하기
                 if (menuNo.equals("1")) {
                     System.out.println("[수정 내용 입력]");
                     System.out.print("제목: ");
@@ -333,32 +86,40 @@ public class BoardApp {
                     System.out.println("보조메뉴: 1.Ok | 2.Cancel");
                     System.out.print("메뉴선택: ");
                     String subMenuNo = scan.nextLine();
+                    //1을 선택해 확정 지었다면 데이터베이스에서 게시물 수정
                     if (subMenuNo.equals("1")) {
                         try {
-                            db.updateBoard(title, content, writer);
+                            db.updateBoard(bno, title, content, writer);
                         }catch (Exception e){
                             e.printStackTrace();
                             break;
                         }
                     }
                 } else if (menuNo.equals("2")) {
+                    //2번 삭제를 선택했다면 방금 읽은 게시물을 삭제하기
                     System.out.println("-------------------------------------------------------------------");
                     System.out.println("보조메뉴: 1.Ok | 2.Cancel");
                     System.out.print("메뉴선택: ");
-                    try {
-                        db.deleteBoard(bno);
-                    }catch (Exception e){
-                        e.printStackTrace();
-                        break;
+                    String subMenuNo = scan.nextLine();
+                    //확정짓는 보조메뉴를 1번 선택했다면 데이터베이스에서 게시물 삭제
+                    if (subMenuNo.equals("1")) {
+                        try {
+                            db.deleteBoard(bno);
+                        } catch (Exception e) {
+                            e.printStackTrace();
+                            break;
+                        }
                     }
                 }
 
             } else if (cmd.equals("3")) {
+                //3번을 선택했다면 게시물 전체 삭제
                 System.out.println("[게시물 전체 삭제]");
                 System.out.println("-------------------------------------------------------------------");
                 System.out.println("보조메뉴: 1.Ok | 2.Cancel");
                 System.out.print("메뉴선택: ");
                 String menuNo = scan.nextLine();
+                //확정짓는 보조메뉴 1번을 선택하면 데이터베이스 전체 게시물 삭제
                 if(menuNo.equals("1")) {
                     try {
                         db.clearBoard();
